Move kernel/api from PythonInterface/inc to kernel/inc and api/inc
- Remove the HEADER_DIR declared in Framework/PythonInterface/CMakeLists.txt
- Include path has been unchanged as files are still in inc/MantidPythonInterface
- Moved ExtractWorkspace to core, as kernel depended on it, and it was in api
