+++ Framework/PythonInterface/CMakeLists.txt
@@ -12,9 +12,6 @@ set(Boost_LIBRARIES ${Boost_LIBRARIES} ${old_boost})
 find_package(Numpy REQUIRED)
 include_directories(SYSTEM ${PYTHON_INCLUDE_PATH} ${PYTHON_NUMPY_INCLUDE_DIR} ${Boost_INCLUDE_DIRS})
 
-# Base of the subproject include dir
-set(HEADER_DIR ${CMAKE_CURRENT_SOURCE_DIR}/inc/MantidPythonInterface)
-
 # Common definitions
 add_definitions(-DBOOST_PYTHON_NO_LIB
                 -DNPY_NO_DEPRECATED_API=NPY_1_7_API_VERSION)
