+++ Framework/PythonInterface/core/inc/MantidPythonInterface/core/ExtractWorkspace.h
