+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/AlgorithmIDProxy.h
