+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/Algorithms/AlgorithmFactoryObserverAdapter.h
