+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/Algorithms/AlgorithmObserverAdapter.h
