+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/AnalysisDataServiceObserverAdapter.h
