+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/BinaryOperations.h
