+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/CloneMatrixWorkspace.h
