+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/ComponentInfoPythonIterator.h
