+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/DetectorInfoPythonIterator.h
