+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/FitFunctions/IFunction1DAdapter.h
