+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/FitFunctions/IFunctionAdapter.h
