+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/FitFunctions/IPeakFunctionAdapter.h
