+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/PythonAlgorithm/DataProcessorAdapter.h
