+++ Framework/PythonInterface/mantid/api/inc/MantidPythonInterface/api/SpectrumInfoPythonIterator.h
