+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/MappingTypeHandler.h
