+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/PropertyManagerFactory.h
