+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/PropertyWithValueFactory.h
