+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/RegisterWorkspacePtrToPython.h
