+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/SequenceTypeHandler.h
