+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/TypeRegistry.h
