+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/Registry/TypedPropertyValueHandler.h
@@ -7,7 +7,7 @@
 #ifndef MANTID_PYTHONINTERFACE_TYPEDPROPERTYVALUEHANDLER_H_
 #define MANTID_PYTHONINTERFACE_TYPEDPROPERTYVALUEHANDLER_H_
 
-#include "MantidPythonInterface/api/ExtractWorkspace.h"
+#include "MantidPythonInterface/core/ExtractWorkspace.h"
 #include "MantidPythonInterface/core/IsNone.h" // includes object.hpp
 #include "MantidPythonInterface/kernel/Registry/PropertyValueHandler.h"
