+++ Framework/PythonInterface/mantid/kernel/inc/MantidPythonInterface/kernel/kernel.h
